Split out Requests and Utils
